client/requirements.txt
kafka-python
requests
flask
mysql-connector-python
uuid